+++ demo/vagrant/client2.hcl
@@ -12,6 +12,10 @@ client {
   # this should be like "nomad.service.consul:4647" and a system
   # like Consul used for service discovery.
   servers = ["127.0.0.1:4647"]
+
+  options {
+    "driver.raw_exec.enable" = "true"
+  }
 }
 
 # Modify our port to avoid a collision with server1 and client1
