+++ nomad/structs/streaming_rpc.go
@@ -4,6 +4,7 @@ import (
 	"errors"
 	"io"
 	"strings"
+	"sync"
 )
 
 // TODO(alexdadgar): move to errors.go
@@ -63,4 +64,24 @@ func (s *StreamingRpcRegistery) GetHandler(method string) (StreamingRpcHandler, 
 	}
 
 	return h, nil
+}
+
+// Bridge is used to just link two connections together and copy traffic
+func Bridge(a, b io.ReadWriteCloser) error {
+	wg := sync.WaitGroup{}
+	wg.Add(2)
+	go func() {
+		defer wg.Done()
+		io.Copy(a, b)
+		a.Close()
+		b.Close()
+	}()
+	go func() {
+		defer wg.Done()
+		io.Copy(b, a)
+		a.Close()
+		b.Close()
+	}()
+	wg.Wait()
+	return nil
 }
